* configure.tgt: Fix i?86-*-linux* entry. 	* rt/sat_arithmetic.c (__hsail_sat_add_u32, __hsail_sat_add_u64, 	__hsail_sat_add_s32, __hsail_sat_add_s64): Use __builtin_add_overflow. 	(__hsail_sat_sub_u8, __hsail_sat_sub_u16): Remove pointless for overflow 	over maximum. 	(__hsail_sat_sub_u32, __hsail_sat_sub_u64, __hsail_sat_sub_s32, 	__hsail_sat_sub_s64): Use __builtin_sub_overflow. 	(__hsail_sat_mul_u32, __hsail_sat_mul_u64, __hsail_sat_mul_s32, 	__hsail_sat_mul_s64): Use __builtin_mul_overflow. 	* rt/arithmetic.c (__hsail_borrow_u32, __hsail_borrow_u64): Use 	__builtin_sub_overflow_p. 	(__hsail_carry_u32, __hsail_carry_u64): Use __builtin_add_overflow_p. 	* rt/misc.c (__hsail_groupbaseptr, __hsail_kernargbaseptr_u64): 	Cast pointers to uintptr_t first before casting to some other integral 	type. 	* rt/segment.c (__hsail_segmentp_private, __hsail_segmentp_group): Likewise. 	* rt/queue.c (__hsail_ldqueuereadindex, __hsail_ldqueuewriteindex, 	__hsail_addqueuewriteindex, __hsail_casqueuewriteindex, 	__hsail_stqueuereadindex, __hsail_stqueuewriteindex): Cast integral value 	to uintptr_t first before casting to pointer. 	* rt/workitems.c (__hsail_alloca_pop_frame): Cast memcpy first argument to 	void * to avoid warning.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@245080 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libhsail-rt/rt/misc.c
@@ -68,8 +68,8 @@ __hsail_debugtrap (uint32_t src, PHSAWorkItem *wi)
 uint32_t
 __hsail_groupbaseptr (PHSAWorkItem *wi)
 {
-  return (uint32_t) (uint64_t) (wi->wg->group_base_ptr
-				- wi->launch_data->group_segment_start_addr);
+  return (uint32_t) (uintptr_t) (wi->wg->group_base_ptr
+				 - wi->launch_data->group_segment_start_addr);
 }
 
 uint64_t
@@ -77,7 +77,7 @@ __hsail_kernargbaseptr_u64 (PHSAWorkItem *wi)
 {
   /* For now assume only a single kernarg allocation at a time.
      Proper kernarg memory management to do.  */
-  return (uint64_t) wi->launch_data->kernarg_addr;
+  return (uint64_t) (uintptr_t) wi->launch_data->kernarg_addr;
 }
 
 uint32_t
